requirements.txt
numpy
pandas
nltk
torch
langchain-community
langchain-core
langchain-huggingface
sentence-transformers
tqdm
python-dotenv
datasets
transformers
ctransformers
hydra-core
omegaconf
ragas